"{#16} {Product} {http://pm.5psolutions.net/public/index.php?path_info=projects/training-andrei-ghiorghe/tasks/16} {This commit made little change in the purpose of a better code structure.}"

+++ database/migrations/2020_08_05_124045_create_order_products_table.php
@@ -17,7 +17,7 @@ class CreateOrderProductsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('order_id')->index();
             $table->unsignedBigInteger('product_id')->index();
-            $table->unsignedBigInteger('quantity');
+            $table->unsignedInteger('quantity');
             $table->nullableTimestamps(0);
         });
     }
